add FAQ.md / fix docstring error / add links to two other textAttack Github sites

+++ textattack/goal_functions/classification/untargeted_classification.py
@@ -13,7 +13,7 @@ class UntargetedClassification(ClassificationGoalFunction):
     the score of the correct label until it is no longer the predicted label.
 
     Args:
-        target_max_score (int): If set, goal is to reduce model output to
+        target_max_score (float): If set, goal is to reduce model output to
             below this score. Otherwise, goal is to change the overall predicted
             class.
     """
